BUG: Fix incorrect message about SlicerRadiomics not installed
If SlicerRadiomics-is installed for the first time with PET-MUST-Segmenter, then upon restart and entering PET-MUST-Segmenter it will warn that SlicerRadiomics is not installed when actually it is installed. It is just importing featureextractor was not successfull because the user has not enter the SlicerRadiomics module yet which installs additional python packages.